[refact] 코드 위치 변경 및 매개변수 수정 #29
