+++ src/main/java/ceos/backend/infra/ses/AwsSESSendMailHandler.java
@@ -2,15 +2,12 @@ package ceos.backend.infra.ses;
 
 import ceos.backend.domain.application.dto.request.CreateApplicationRequest;
 import ceos.backend.global.common.dto.AwsSESMail;
-import ceos.backend.global.common.dto.mail.*;
+import ceos.backend.global.common.dto.AwsSESPasswordMail;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
 import org.thymeleaf.context.Context;
-
-import java.util.ArrayList;
-import java.util.List;
 
 @Component
 @RequiredArgsConstructor
@@ -29,5 +26,13 @@ public class AwsSESSendMailHandler {
                                                         .getGeneration());
         final Context CONTEXT = awsSESMailGenerator.generateApplicationMailContext(awsSESMail);
         awsSesUtils.singleEmailRequest(TO, SUBJECT, "sendApplicationMail", CONTEXT);
+    }
+
+    @EventListener(AwsSESPasswordMail.class)
+    public void handle(AwsSESPasswordMail awsSESPasswordMail) {
+        final String TO = awsSESPasswordMail.getEmail();
+        final String SUBJECT = awsSESMailGenerator.generatePasswordMailSubject();
+        final Context CONTEXT = awsSESMailGenerator.generatePasswordMailContext(awsSESPasswordMail);
+        awsSesUtils.singleEmailRequest(TO, SUBJECT, "sendPasswordMail", CONTEXT);
     }
 }
